rust: Remove client form ws_services example
